Add GL::Texture
I think I should make this whole thing header only because it means that the client only uses what they need. If the client wants to use the OpenGL part of the library, they link against OpenGL and include the appropriate headers. If the client doesn't want to use OpenGL, they don't have to link against OpenGL.

The only issue with that is that I have to put the code into a translation unit for the compiler to see it and check for errors.