Add B2::CollisionPairs
I hope I can reuse all of this stuff. My next game won't actually use Box2D so I won't be able to test it out immediately. While on the topic of the next game...

I've found that the physics engine restricts the puzzles that can be created. When making the levels, I've found myself fighting the physics engine. There are some puzzles that I wanted to create but weren't possible because Box2D was getting in the way. Physics also makes it more difficult for the player to move around. Moving horizontally is easy but moving vertically is difficult because of gravity. When physics are involved, the game becomes more about "how do I jump over there?" than "where should I jump?".

I will learn from all of the mistakes I made in the creation of Buttons when making a new game. Perhaps I'll call it "The Machine". Now that sounds cool!