Use a std::unique_ptr<char []> as the storage of ComposeString
This way, we have more control over the memory which the string is contained in. We can repeatedly insert characters into the string and then change the size of the string rather than calling std::string::push_back. push_back checks if capacity needs to be changed, copies the character and increments size.

This implementation also makes writing numbers more efficient. std::to_chars can write directly to the string and the size can be increased after. This is much faster than std::to_string.