Rename QuadAllocator to JobManager
JobManager is an incredible class. It's an efficient layer on top of the Renderer. Quads and rendering params are stored in the JobManager and then rendered using the Renderer.

This is a huge improvement over The Machine which had all of the functionality of the JobManager in the Render and Jobs weren't just data associated with IDs, Jobs were classes!