Name faster quad writer FastQuadWriter
I think I'm going keep these too implementations together until I can clean up the implementation of FastQuadWriter. The interfaces are the same so I could just delete the slow one and use the fast one but implementation is just as important as the interface. I have a nice clean interface and I want a nice clean implementation as well. The original QuadWriter has a clean implementation and FastQuadWriter doesn't.

But for the sort of things I'm using these classes for, QuadWriter is fast enough.