Delete G2D::FastQuadWriter
I did some benchmarks with Classic Tower Defence. This is the average time in milliseconds it takes to push everything into the QuadWriter (it's tabulated in a monospace font).

                    DEBUG          RELEASE
QuadWriter          0.166         0.0369
FastQuadWriter      0.154          0.0292

FastQuadWriter is faster than QuadWriter but not by much. Using a pointer to the back is just a little fast than using std::vector::back(). QuadWriter is already way faster than it needs to be. Classic Tower Defence is running at 20fps so we have to get the quads pushed within 50ms. The standard QuadWriter could push all of the quads 1355 times in 50ms. There's just no need to have a more complicated but slightly faster QuadWriter.