POINTERS!!!
I'm not exposing pointers in Stela!

Damn, isn't Stela a nice name! It's short. It's unique. It's kind of like Perl. Pearl is a real word but Perl is spelt different. So when you google "pearl" you get white spheres and if you google "perl" you get a programming language. Same with Stela. Stela is also an acronym which makes it extra awesome.

The future is bright. :-)