Refactor Grid transform
I was originally going to write it like this but I thought the template solution would be faster because it sort of works like expression templates. I did some playing around with compiler explorer and the simple solution is either just as fast or faster than the template solution. I guess simple code is easier for the compiler to understand which makes it easier to optimise.