Rough draft of G2D::Sequence
I want this to be able to do fade-to-black between levels like The Machine. I also want if to be able to do things like the intro to The Machine. The intro to The Machine was baked onto a texture and the frames were rendered. The texture was 8192! I want to be able to compress that down to a few sprites that move around the screen.